requirements.txt
fastapi
uvicorn
ultralytics
torch
torchvision
pillow
numpy
